requirements.txt
cython
nltk
numpy
chumpy
requests

dv
opencv-python
matplotlib
imageio
imageio-ffmpeg
pyyaml
tqdm
boto3
scipy
yacs